+++ outputs/x86_64-linux/tests/bun/expected.nix
@@ -1,0 +1,11 @@
+{
+  lib,
+}:
+let
+  hosts = [
+    "apollo"
+    "athena"
+    "generic"
+  ];
+in
+lib.genAttrs hosts (_: "1.3.14")