+++ outputs/x86_64-linux/tests/bun/expr.nix
@@ -1,0 +1,15 @@
+{
+  lib,
+  outputs,
+}:
+let
+  hosts = [
+    "apollo"
+    "athena"
+    "generic"
+  ];
+in
+lib.genAttrs hosts (
+  name:
+  outputs.nixosConfigurations.${name}.pkgs.lib.getVersion outputs.nixosConfigurations.${name}.pkgs.bun
+)